Soft reset to resolve conflicts
