+++ frontend/src/Components/Events/eventstore.js
@@ -20,3 +20,29 @@ export async function getEventActionButtons() {
         return await response.text();
 }
 
+export async function requestAction(event, action, userActedOn) {
+    const response = await fetch(`/api/eventactionrecords/`, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+            "X-CSRFToken": document.cookie
+                .split("; ")
+                .find((element) => element.startsWith("csrftoken="))
+                .split("=")[1],
+        },
+        body: JSON.stringify({
+            event: event.pk,
+            acted_on: userActedOn.user_id,
+            action: action,
+            extra_data: "{}",
+            points: action === "Check Off" ? event.points: 0,
+        }),
+    });
+    if (response.status == 200 || response.status == 201) {
+        const result = await response.json();
+        window.location.reload();
+    } else {
+        const message = await response.json();
+        alert(`${response.statusText}: ${message['detail']}`);
+    }
+}
