+++ myapp/api/views.py
@@ -77,6 +77,88 @@ load_dotenv(os.path.join(BASE_DIR, '.env'))
 #################################################################
 ## View Sets
 #################################################################
+class EventViewSet(ModelViewSet):
+    def get_serializer_class(self):
+        if self.request.method in SAFE_METHODS:
+            return EventGetSerializer
+        return EventPostSerializer
+
+    def get_permissions(self):
+        if self.request.method not in SAFE_METHODS:
+            permission_classes = [HasAdminPermissions]
+        else:
+            permission_classes = [IsAuthenticated]
+        return [permission() for permission in permission_classes]
+
+    def get_queryset(self):
+        if is_admin(self.request.user):
+            return Event.objects.all()
+
+        user_groups = self.request.user.groups.all()
+
+        if self.request.method in SAFE_METHODS:
+            permitted_events_attr = "viewable_events"
+            viewable_posts = Event.objects.all().filter(anon_viewable=True)
+
+        for group in user_groups:
+            viewable_posts = viewable_posts | getattr(group, permitted_events_attr).all()
+
+        if not is_admin(self.request.user):
+            viewable_posts = viewable_posts.filter(is_draft=False)
+
+        return viewable_posts.distinct()
+
+    @action(detail=True, methods=["get"])
+    def relevant_users(self, request, pk):
+        if not is_admin(self.request.user):
+            return Response([])
+        else:
+            relevant_users = CustomUser.objects.filter(
+                actions_received__event__pk=pk
+            ).distinct()
+        serializer = UserSerializer(relevant_users, many=True)
+        if serializer.is_valid:
+            return Response(serializer.data)
+
+class EventActionRecordViewSet(ModelViewSet):
+    serializer_class = EventActionRecordGetSerializer
+    queryset = EventActionRecord.objects.all()
+    
+    def get_permissions(self):
+        permission_classes = [IsAuthenticated]
+        return [permission() for permission in permission_classes]
+
+    def get_queryset(self):
+        user = self.request.user
+        if is_admin(user):
+            return super().get_queryset()
+
+        return super().get_queryset().filter(user=self.request.user)
+
+    def get_serializer_class(self):
+        if self.request.method in SAFE_METHODS:
+            return EventActionRecordGetSerializer
+        else:
+            return EventActionRecordPostSerializer
+
+    @action(detail=False)
+    def get_record_of_action(self, request, action):
+        return self.queryset.filter(action=action)
+
+    def create(self, request, *args, **kwargs):
+        try:
+            serializer = EventActionRecordPostSerializer(data=request.data)
+            if serializer.is_valid():
+                action = serializer.data["action"]
+                event_action.all[action](request, serializer.data)
+
+                return super().create(request, *args, **kwargs)
+        except Exception as e:
+            return Response({'detail': str(e)}, status=status.HTTP_400_BAD_REQUEST)
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
 class UserViewSet(ReadOnlyModelViewSet):
     queryset = CustomUser.objects.all()
     serializer_class = UserSerializer
@@ -120,6 +202,10 @@ class UserProfileView(APIView):
             return Response(serializer_data, status=status.HTTP_200_OK)
 
 # Note: Making both of these read only so they can't be edited directly from the portal
+class EventTypeViewSet(ReadOnlyModelViewSet):
+    queryset = EventType.objects.all()
+    serializer_class = EventTypeSerializer
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
 class GroupsViewSet(ReadOnlyModelViewSet):
     queryset = Group.objects.all()
@@ -130,6 +216,56 @@ class GroupsViewSet(ReadOnlyModelViewSet):
 ## Specific Views for GET Requests
 #################################################################
 
+@api_view(["GET"])
+def EventActionRecordsForEventUserPair(request, event_pk, other_user_id):
+    serializer = EventActionRecordGetSerializer(
+        EventActionRecord.objects.filter(
+            event__pk=event_pk, acted_on__user_id=other_user_id
+        ),
+        many=True,
+    )
+    if serializer.is_valid:
+        return Response(serializer.data)
+
+    raise act_exceptions.ForbiddenException
+
+@api_view(["GET"])
+def EventActionView(request):
+    permitted_self_actions = []
+    permitted_other_actions = []
+    for action in event_action.self_actions.keys():
+        if (
+            action not in event_action.eventless_actions.keys()
+        ):
+            permitted_self_actions.append(action)
+
+    for action in event_action.other_actions.keys():
+        if (
+            is_admin(request.user)
+            and action not in event_action.eventless_actions.keys()
+        ):
+            permitted_other_actions.append(action)
+
+    return Response(
+        {
+            "self_actions": permitted_self_actions,
+            "other_actions": permitted_other_actions,
+        }
+    )
+
+
+@api_view(["GET"])
+def EventlessActionView(request):
+    permitted_eventless_actions = []
+    for action in event_action.eventless_actons.keys():
+        if request.user.has_perm(f"can_{action.lower().replace(' ', '_')}"):
+            permitted_eventless_actions.append(action)
+
+    return Response(
+        {
+            "actions": permitted_eventless_actions,
+        }
+    )
 
 @api_view(["GET"])
 def PermissionsView(request):
