--- myapp/api/views/user_views.py
@@ -1,1 +1,0 @@
-# TODO: Move user-related view code here
